requirements.txt
astor==0.8.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
funcsigs==1.0.2
gast==0.3.3
graphviz==0.14.2
idna==2.10
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.2
nltk==3.5
numpy==1.19.4
objgraph==3.5.0
opencv-python==4.4.0.46
paddlepaddle==1.8.5
pathlib==1.0.1
Pillow==8.0.1
prettytable==1.0.1
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
rarfile==4.0
regex==2020.10.28
requests==2.24.0
scipy==1.5.3
six==1.15.0
tqdm==4.51.0
urllib3==1.25.11
wcwidth==0.2.5
